Anpassung Tab-Ansicht, weitere Bilder bei den Newseinträge, Div. Anpassungen

+++ src/app/sqlite.service.ts
@@ -1,6 +1,7 @@
 import {Injectable} from '@angular/core';
 import {SQLite, SQLiteObject, SQLiteDatabaseConfig} from '@ionic-native/sqlite/ngx';
 import {Platform} from '@ionic/angular';
+import {Device} from '@ionic-native/device/ngx';
 
 @Injectable({
     providedIn: 'root'
@@ -11,7 +12,8 @@ export class SqliteService {
 
     constructor(
         private sqLite: SQLite,
-        private platform: Platform
+        private platform: Platform,
+        private device: Device
     ) {
     }
 
@@ -28,19 +30,21 @@ export class SqliteService {
             };
         }
 
-        this.sqLite.create(this.dbConfig).then((db: SQLiteObject) => {
-            /*db.executeSql('DROP TABLE IF EXISTS api_content', [])
-                .then(() => console.log('Drop table if exists content'))
-                .catch(e => console.log(JSON.stringify(e)));*/
+        if (this.device.cordova != null && (this.platform.is('android') || this.platform.is('ios'))) {
+            this.sqLite.create(this.dbConfig).then((db: SQLiteObject) => {
+                /*db.executeSql('DROP TABLE IF EXISTS api_content', [])
+                        .then(() => console.log('Drop table if exists content'))
+                        .catch(e => console.log(JSON.stringify(e)));*/
 
-            console.log('Database created.');
+                console.log('Database created.');
 
-            db.executeSql('CREATE TABLE IF NOT EXISTS api_content (content_id INTEGER PRIMARY KEY, api_path TEXT, api_data TEXT)')
-                .then(() => {
-                })
-                .catch(e => console.log(e));
+                db.executeSql('CREATE TABLE IF NOT EXISTS api_content (content_id INTEGER PRIMARY KEY, api_path TEXT, api_data TEXT)')
+                    .then(() => {
+                    })
+                    .catch(e => console.log(e));
 
-            this.db = db;
-        }).catch(e => console.log(e));
+                this.db = db;
+            }).catch(e => console.log(e));
+        }
     }
 }
